api_yamdb/requirements.txt
atomicwrites==1.4.1
attrs==22.2.0
asgiref==3.2.10
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==2.2.16
django-extensions==3.1.5
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
idna==3.4
iniconfig==1.1.1
packaging==22.0
pluggy==0.13.1
psycopg2-binary==2.9.4
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
pytz==2022.7
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.13
